% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRM_R.R
\name{LRM_R}
\alias{LRM_R}
\title{Linear Regression Model using R}
\usage{
LRM_R(formula, x = FALSE, y = FALSE)
}
\arguments{
\item{formula}{a symbolic description of the model to be fitted.}

\item{x}{logicals.It is TRUE after the model matrix returned.}

\item{y}{logicals.It is TRUE after the reponse matrix returned.}
}
\value{
LRM_R returns a set of simple and multiple linear regression result of formula.
An object of class "LRM_R" is a list containing at least the following components:
\itemize{
  \item coefficients -  The estimated value, standard deviation and statistics value of coefficients
  \item r.squared - The coefficient of determination of the fitted model
  \item fstatistic - The result of F-statistic of the fitted model
  \item Y_model - The result of linear regression for response matrix
}
}
\description{
LRM_R is used to implementing simple and multiple linear regression using least squares method in R.
It can carry out the estimated value，standard deviation and statistics value of coefficients
}
\examples{
#Using R's default dataset
LRM_R(mtcars$mpg~mtcars$disp +mtcars$wt+mtcars$qsec )

#In addition to data frames, it also supply vectors.
a <- c(1.1,2.5,4.1,5.0,6.5,2.6,1.1,3.5,3.3,4.1)
b <- c(4.8,4.1,4.4,3.5,5.8,3.8,6.0,4.8,4.3,4.6)
LRM_R(a~b)$coefficients
LRM_R(a~b)$r.squared


}
